:sparkles: model created for talk :recycle: getTalks function updated with the model
